+++ Naheulbook.Data/Models/JobRequirement.cs
@@ -2,7 +2,7 @@ namespace Naheulbook.Data.Models
 {
     public class JobRequirement
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public long? MinValue { get; set; }
         public long? MaxValue { get; set; }
@@ -10,7 +10,7 @@ namespace Naheulbook.Data.Models
         public string StatName { get; set; }
         public Stat Stat { get; set; }
 
-        public long JobId { get; set; }
+        public int JobId { get; set; }
         public Job Job { get; set; }
     }
 }
