+++ Naheulbook.Data/Models/OriginSkill.cs
@@ -4,10 +4,10 @@ namespace Naheulbook.Data.Models
     {
         public bool Default { get; set; }
 
-        public long OriginId { get; set; }
+        public int OriginId { get; set; }
         public Origin Origin { get; set; }
 
-        public long SkillId { get; set; }
+        public int SkillId { get; set; }
         public Skill Skill { get; set; }
     }
 }
