+++ Naheulbook.Data/Models/SkillEffect.cs
@@ -2,9 +2,9 @@
 {
     public class SkillEffect
     {
-        public long SkillId { get; set; }
+        public int SkillId { get; set; }
         public string StatName { get; set; }
-        public long Value { get; set; }
+        public int Value { get; set; }
         public Skill Skill { get; set; }
     }
 }
