+++ Naheulbook.Data/Models/SpecialityModifier.cs
@@ -2,11 +2,11 @@ namespace Naheulbook.Data.Models
 {
     public class SpecialityModifier
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Stat { get; set; }
-        public long Value { get; set; }
+        public int Value { get; set; }
 
-        public long SpecialityId { get; set; }
+        public int SpecialityId { get; set; }
         public virtual Speciality Speciality { get; set; }
     }
 }
